🔨 allow cancel friend request if user not found

+++ firebase/functions/src/callables/cancelOutgoingFriendRequest.ts
@@ -1,7 +1,7 @@
 import * as functions from 'firebase-functions'
 import admin = require('firebase-admin')
 import { firestore } from 'firebase-admin'
-import { getFriendRequestsDocShortcut, getUserDataByUid, getUserDataByUsername  } from '../helpers'
+import { getFriendRequestsDocShortcut, getUserDataByUid, getUserDataByUsername } from '../helpers'
 
 const db = admin.firestore()
 
@@ -25,7 +25,6 @@ exports.cancelOutgoingFriendRequest = functions.https.onCall((data, context) => 
     const [senderUsername, recipientUid] = await Promise.all([getSenderUsername(), getRecipientUid()])
 
     if (!senderUsername) throw new functions.https.HttpsError('aborted', 'Sender user could not be found')
-    if (!recipientUid) throw new functions.https.HttpsError('aborted', 'Recipient user could not be found')
 
     // remove the outgoing request from the sender's /userData/friendRequests/outgoing field
     transaction.set(friendRequestsDoc(senderUid), {
@@ -34,11 +33,13 @@ exports.cancelOutgoingFriendRequest = functions.https.onCall((data, context) => 
       }
     }, { merge: true })
 
-    // remove the incoming request from the recipient's /userData/friendRequests/incoming field
-    transaction.set(friendRequestsDoc(recipientUid), {
-      incoming: {
-        [senderUsername]: firestore.FieldValue.delete()
-      }
-    }, { merge: true })
+    if (recipientUid) {
+      // remove the incoming request from the recipient's /userData/friendRequests/incoming field
+      transaction.set(friendRequestsDoc(recipientUid), {
+        incoming: {
+          [senderUsername]: firestore.FieldValue.delete()
+        }
+      }, { merge: true })
+    }
   })
 })
